refactor(slice-45): split frontend components into theme folders
Move screens, chrome, experiment, and stats into dedicated packages
with co-located tests so App imports match the theme map layout.
