+++ CLAUDE.md
@@ -153,14 +153,13 @@ List/detail: dashboard or `GET /experiments` / `GET /experiments/{id}` (see `htt
 | `frontend/src/components/PollingIndicator.tsx` | Subtle "Syncing..." badge during background polls |
 | `frontend/src/components/ConfirmDeleteModal.tsx` | Delete confirmation modal with experiment details and stats |
 | `frontend/src/components/ExperimentControlButtons.tsx` | Pause, resume, cancel buttons on detail screen |
+| `frontend/src/components/screens/` | Feature screens (`Experiments`, `Detail`, `SearchExplorer`) + co-located tests |
+| `frontend/src/components/chrome/` | Shell chrome (`DashboardShell`, `AppPageChrome`, `Pagination`, …) |
+| `frontend/src/components/experiment/` | Experiment controls/progress/modals + detail chrome |
+| `frontend/src/components/stats/` | Vector DB stats panels + `StatTile`/`StatRow` |
 | `frontend/src/hooks/useExperimentDetail.ts` | Detail hydrate/poll/db-stats controller (Slice 45) |
-| `frontend/src/components/experimentDetail/` | Detail presentational chrome (StatusBadge, StatCard, …) |
 | `frontend/src/components/explore/ExplorePanels.tsx` | Search Explorer tabs + sidebar panels |
 | `frontend/src/components/chrome/Pagination.tsx` | Shared list/detail/explorer pagination chrome (Slice 45) |
-| `frontend/src/components/stats/StatTile.tsx` | Shared vector-DB metric tile (+ `StatRow`) |
-| `frontend/src/components/ExperimentsScreen.tsx` | Experiments list with collapsible rows, vector DB stats, delete |
-| `frontend/src/components/ExperimentDetailScreen.tsx` | Detail view with overview metrics, outcome banners, runs table |
-| `frontend/src/components/VectorDbStatsPanel.tsx` | Cluster-grouped storage stats panel |
 | `frontend/src/components/CollapsibleCard.tsx` | Reusable collapsible section (localStorage persistence) |
 | `frontend/src/utils/experimentStatus.ts` | Run outcome summarization + terminal status helpers |
 | `frontend/src/utils/completionReason.ts` | Shared `completion_reason` → label mapping |
