+++ docs/plan/DECISIONS.md
@@ -157,4 +157,4 @@
 | 151 | 2026-07-28 | 45 | Land `server/core/rerank/` (`reranker`, `local_reranker`); `sys.modules` shims at old paths | Completes core Feature provider cluster from theme map; search imports canonical `server.core.rerank.reranker` | Fold into #150 / leave flat |
 | 152 | 2026-07-28 | 45 | Land `server/db/{ports,mongo,postgres}/`; shims at old flat `server.db.*` paths except **no** `server/db/postgres.py` (collides with package); pool API re-exported on `server.db.postgres` package; `schema.sql` co-located with `postgres.py` | Must hotspot #2; preserves Protocols + factory; `from server.db import postgres` / `close_pool` stays valid | Split db across multiple PRs / invent `pg/` package rename |
 | 153 | 2026-07-28 | 45 | Fat-surface slim: `experiments_lifecycle.py` + `cli/display.py`; keep routers/commands as façades | Execution order step 3 after db; pre→post lines experiments 553→388, cli/main 519→246 | Defer slim until FE primitives / leave god modules |
-| 156 | 2026-07-28 | 45 | FE screen SLAP: `useExperimentDetail` + `experimentDetail/*` chrome; `experimentList/labels`; `explore/ExplorePanels`; duration helpers in `experimentDetailProgress` | Step 6 before folder move; pre→post Detail 1615→986, Experiments 764→670, Explorer 1131→426; FE floors still green (261 tests) | Big-bang folder move without extracts / leave god screens |
+| 157 | 2026-07-28 | 45 | Split `frontend/src/components/` into `screens/`, `chrome/`, `experiment/`, `stats/` (+ co-located tests); nest `experimentDetail` under `experiment/` | Step 7 after SLAP; App + mocks retargeted; FE floors still green | Leave flat components/ / move without co-located tests |
