+++ frontend/src/components/chrome/PollingIndicator.test.tsx
