+++ frontend/src/components/chrome/PollingIndicator.tsx
