+++ frontend/src/components/experiment/experimentDetail/DimensionBadge.tsx
