+++ frontend/src/components/experiment/experimentDetail/StatCard.tsx
