+++ frontend/src/components/experiment/experimentDetailProgress.ts
